requirements.txt
tensorflow==2.12.0
remotezip==0.12.1
tqdm==4.65.0
opencv_python==4.7.0.72
tf-models-official